pyproject.toml
[project]
name = "xfads"
version = "0.7.1"
description = "eXponential Family Dynamical Systems"
authors = []
dependencies = [
    "torch==2.2.2; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "torch>=2.6; sys_platform != 'darwin' or platform_machine != 'x86_64'",
    "numpy>=1.26.4",
    "scipy>=1.13.1",
    "tqdm>=4.66.4",
    "scikit-learn>=1.5.0",
    "lightning>=2.2.5",
    "matplotlib>=3.9.2",
    "hydra-core>=1.3.2",
    "einops>=0.8.0",
    "pyyaml>=6.0.1",
    "sympy>=1.12",
    "torchvision>=0.2.2",
    "joblib>=1.4.2",
    "skorch==1.0.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["xfads"]
